refactor: get projects from `allProjects`
